Add styles to publicDecks
